Add support for customizing the user's group
In order to allow users of this lwrp to set the group when rendering the
zsh template, this commit adds group to the resource definition and
makes the render_zshrc method accept group as an argument. In order to
preserve backwards compatibility it also allows group to not be set and
default to the previous behavior of using the username as the group.